Pagination in Android
Pagination is the combination of Android Recycler view and the PHP backend. As we know in PHP backend we can perform this operation in 2 or 3 ways as per based on sql quaries like by using BETWEEN, LIMIT and by AND quaries. so choose with is comfortable for you but on Android there is only this kindd approach will work.